requirements.txt
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
numpy==2.0.2
pandas==2.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
soupsieve==2.7
typing_extensions==4.14.0
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.4.0
